Replace Lib.RemoteSet with less abstract GameSet
I'll try to remember to make things abstract IF it's actually useful.

GameSet handles batch loading of games, and keeps track of their states.

This is still WIP, because if there are temporary errors, they are not yet
retried. Also it seems this might not detect missing games correctly yet.